增加校验工具以及后台校验拦截

+++ server/api/v1/sys_api.go
@@ -7,6 +7,7 @@ import (
 	"gin-vue-admin/model/request"
 	resp "gin-vue-admin/model/response"
 	"gin-vue-admin/service"
+	"gin-vue-admin/utils"
 	"github.com/gin-gonic/gin"
 )
 
@@ -21,6 +22,17 @@ import (
 func CreateApi(c *gin.Context) {
 	var api model.SysApi
 	_ = c.ShouldBindJSON(&api)
+	ApiVerify := utils.Rules{
+		"Path":      {utils.NotEmpty()},
+		"Description":      {utils.NotEmpty()},
+		"ApiGroup":      {utils.NotEmpty()},
+		"Method":      {utils.NotEmpty()},
+	}
+	ApiVerifyErr := utils.Verify(api, ApiVerify)
+	if ApiVerifyErr!=nil {
+		response.FailWithMessage(ApiVerifyErr.Error(), c)
+		return
+	}
 	err := service.CreateApi(api)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("创建失败，%v", err), c)
@@ -40,6 +52,14 @@ func CreateApi(c *gin.Context) {
 func DeleteApi(c *gin.Context) {
 	var a model.SysApi
 	_ = c.ShouldBindJSON(&a)
+	ApiVerify := utils.Rules{
+		"ID":      {utils.NotEmpty()},
+	}
+	ApiVerifyErr := utils.Verify(a.Model, ApiVerify)
+	if ApiVerifyErr!=nil {
+		response.FailWithMessage(ApiVerifyErr.Error(), c)
+		return
+	}
 	err := service.DeleteApi(a)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("删除失败，%v", err), c)
@@ -62,6 +82,15 @@ func GetApiList(c *gin.Context) {
 	// 此结构体仅本方法使用
 	var sp request.SearchApiParams
 	_ = c.ShouldBindJSON(&sp)
+	ApiVerify := utils.Rules{
+		"Page": {utils.NotEmpty()},
+		"PageSize": {utils.NotEmpty()},
+	}
+	ApiVerifyErr := utils.Verify(sp.PageInfo, ApiVerify)
+	if ApiVerifyErr!=nil {
+		response.FailWithMessage(ApiVerifyErr.Error(), c)
+		return
+	}
 	err, list, total := service.GetAPIInfoList(sp.SysApi, sp.PageInfo, sp.OrderKey, sp.Desc)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("获取数据失败，%v", err), c)
@@ -86,6 +115,14 @@ func GetApiList(c *gin.Context) {
 func GetApiById(c *gin.Context) {
 	var idInfo request.GetById
 	_ = c.ShouldBindJSON(&idInfo)
+	ApiVerify := utils.Rules{
+		"Id": {utils.NotEmpty()},
+	}
+	ApiVerifyErr := utils.Verify(idInfo, ApiVerify)
+	if ApiVerifyErr!=nil {
+		response.FailWithMessage(ApiVerifyErr.Error(), c)
+		return
+	}
 	err, api := service.GetApiById(idInfo.Id)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("获取数据失败，%v", err), c)
@@ -105,6 +142,17 @@ func GetApiById(c *gin.Context) {
 func UpdateApi(c *gin.Context) {
 	var api model.SysApi
 	_ = c.ShouldBindJSON(&api)
+	ApiVerify := utils.Rules{
+		"Path":      {utils.NotEmpty()},
+		"Description":      {utils.NotEmpty()},
+		"ApiGroup":      {utils.NotEmpty()},
+		"Method":      {utils.NotEmpty()},
+	}
+	ApiVerifyErr := utils.Verify(api, ApiVerify)
+	if ApiVerifyErr!=nil {
+		response.FailWithMessage(ApiVerifyErr.Error(), c)
+		return
+	}
 	err := service.UpdateApi(api)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("修改数据失败，%v", err), c)
